start encapsulating code editor
